requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
mysql-connector-python==9.3.0
narwhals==1.41.0
networkx==3.5
numpy==2.2.6
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scipy==1.15.3
sendgrid==6.10.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
starkbank-ecdsa==2.2.0
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.67.1
twilio==9.6.2
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.151
ultralytics-thop==2.0.14
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
